Move useWorkspaceTabInfo into a side-effect-free module (BL-15971)
Devin observed (informationally, PR #8185) that CollectionsTabBookPane
importing useWorkspaceTabInfo from TopBar.tsx also runs TopBar's
module-scope WireUpForWinforms registration in the collection bundle.
Harmless today, but fragile; Andrew opted to extract the hook. The
workspace-tab types, defaultWorkspaceTabState, getActiveWorkspaceTab,
and the hook now live in workspaceTabInfo.ts, which has no side
effects; TopBar.tsx and all other consumers import from it.

Also rides along: a PAPERCUTS.md entry about the NoErrorForEmptyAudio
test being fragile against temp paths containing "17" (an environmental
failure observed during preflight, unrelated to this branch).

Tests: pnpm typecheck passed; pnpm lint 0 errors; full Vitest suite
674 passed; isolated production bundle build (build/agent-vite.sh)
succeeded.
